ShinRok-master/model/requirements.txt
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
PyMySQL==1.4.6
boto3==1.34.124
python-dotenv==1.0.1
torch==2.2.0+cu121
torchvision==0.17.0+cu121
Pillow==10.3.0
numpy==1.26.4
requests==2.32.3
quart==0.19.6
aiohttp==3.9.5

# cd model
# pip install -r requirements.txt